fix(tracer): fixed create exit span with context error when current context not to be sampled
In the tracer CreateExitSpanWithContext function, when the context not to be sample in this time, it
will be return a "span type is wrong" error.

Because trying to check the span is NoopSpan when after called CreateLocalSpan function, the convertion is expected a struct value.

But actual, the CreateLocalSpan function return a Span is a interface type, a interface type is a pointer in golang.

So we need to change the convertion to expect a pointer of NoopSpan type.

fix #113